+++ src/pages/Register.jsx
@@ -1,0 +1,7 @@
+// import "./App.css";
+
+function Register() {
+  return <div> 빈페이지</div>;
+}
+
+export default Register;
